+++ src/App.js
@@ -13,6 +13,7 @@ import {
 // import Admin from "./pages/admin/Admin";
 import Home from "./pages/home/Home";
 import Profile from "./pages/profile/Profile";
+import Login from "./modules/login/Login";
 
 // import logo from "./logo.svg";
 // import "./App.css";
@@ -27,7 +28,7 @@ export default function App() {
           <Routes>
             <Route exact path="/" element={<Home />} />
             <Route exact path="/home" element={<Home />} />
-            <Route exact path="/login" element={<LoginPage />} />
+            <Route exact path="/login" element={<Login />} />
 
             <Route
               exact
@@ -45,31 +46,24 @@ export default function App() {
   );
 }
 
-const fakeAuth = {
-  isAuthenticated: false,
-  signin(cb) {
-    fakeAuth.isAuthenticated = true;
-    setTimeout(cb, 100); // fake async
-  },
-  signout(cb) {
-    fakeAuth.isAuthenticated = false;
-    setTimeout(cb, 100);
-  },
-};
+// A wrapper for <Route> that redirects to the login
+// screen if you're not yet authenticated.
 
-/** For more details on
- * `authContext`, `ProvideAuth`, `useAuth` and `useProvideAuth`
- * refer to: https://usehooks.com/useAuth/
- */
 const authContext = createContext();
+
+function PrivateRoute({ children, ...rest }) {
+  let auth = useContext(authContext);
+  let location = useLocation();
+  if (!auth.user) {
+    return <Navigate to="/login" state={{ from: location }} replace />;
+  }
+
+  return children;
+}
 
 function ProvideAuth({ children }) {
   const auth = useProvideAuth();
   return <authContext.Provider value={auth}>{children}</authContext.Provider>;
-}
-
-function useAuth() {
-  return useContext(authContext);
 }
 
 function useProvideAuth() {
@@ -96,38 +90,17 @@ function useProvideAuth() {
   };
 }
 
-function AuthButton() {
-  let history = useNavigate();
-  let auth = useAuth();
-
-  return auth.user ? (
-    <p>
-      Welcome!{" "}
-      <button
-        onClick={() => {
-          auth.signout(() => history("/"));
-        }}
-      >
-        Sign out
-      </button>
-    </p>
-  ) : (
-    <p>You are not logged in.</p>
-  );
-}
-
-// A wrapper for <Route> that redirects to the login
-// screen if you're not yet authenticated.
-
-function PrivateRoute({ children, ...rest }) {
-  let auth = useAuth();
-  let location = useLocation();
-  if (!auth.user) {
-    return <Navigate to="/login" state={{ from: location }} replace />;
-  }
-
-  return children;
-}
+const fakeAuth = {
+  isAuthenticated: false,
+  signin(cb) {
+    fakeAuth.isAuthenticated = true;
+    setTimeout(cb, 100); // fake async
+  },
+  signout(cb) {
+    fakeAuth.isAuthenticated = false;
+    setTimeout(cb, 100);
+  },
+};
 
 // const PrivateRoute = () => {
 //   const statusLogin = false;
@@ -145,24 +118,3 @@ function PrivateRoute({ children, ...rest }) {
 // function ProtectedPage() {
 //   return <h3>Protected</h3>;
 // }
-
-function LoginPage() {
-  let history = useNavigate();
-  let location = useLocation();
-  let auth = useAuth();
-
-  let { from } = location.state || { from: { pathname: "/" } };
-  let login = () => {
-    auth.signin(() => {
-      // history.replace(from);
-      history(from, { replace: true });
-    });
-  };
-
-  return (
-    <div>
-      <p>You must log in to view the page at {from.pathname}</p>
-      <button onClick={login}>Log in</button>
-    </div>
-  );
-}
